+++ packages/agent-infra/mcp-client/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @agent-infra/mcp-client
 
+## 1.1.6
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [a85ef50]
+  - @agent-infra/mcp-shared@1.1.6
+
+## 1.1.6-beta.10
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.10
+
+## 1.1.6-beta.9
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.9
+
+## 1.1.6-beta.8
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.8
+
+## 1.1.6-beta.7
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.7
+
+## 1.1.6-beta.6
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.6
+
+## 1.1.6-beta.5
+
+### Patch Changes
+
+- @agent-infra/mcp-shared@1.1.6-beta.5
+
 ## 1.1.6-beta.4
 
 ### Patch Changes