+++ packages/agent-infra/mcp-servers/commands/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## 0.4.0 2024-12-10 - add logging
 
+## 1.1.6
+
+### Patch Changes
+
+- a85ef50: chore: publish
+
+## 1.1.6-beta.10
+
+## 1.1.6-beta.9
+
+## 1.1.6-beta.8
+
+## 1.1.6-beta.7
+
+## 1.1.6-beta.6
+
+## 1.1.6-beta.5
+
 ## 1.1.6-beta.4
 
 ## 1.1.6-beta.3