+++ packages/agent-infra/search/search/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @agent-infra/search
 
+## 0.0.3
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [eb293fb]
+- Updated dependencies [a85ef50]
+- Updated dependencies [eb293fb]
+  - @agent-infra/shared@0.0.2
+  - @agent-infra/browser-search@0.0.3
+
 ## 0.0.3-beta.3
 
 ### Patch Changes