+++ packages/ui-tars/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ui-tars/cli
 
+## 1.2.2
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [a85ef50]
+  - @ui-tars/operator-nut-js@1.2.2
+  - @ui-tars/operator-adb@1.2.2
+  - @ui-tars/sdk@1.2.2
+
 ## 1.2.2-beta.3
 
 ### Patch Changes