+++ packages/ui-tars/operators/browser-operator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ui-tars/operator-browser
 
+## 1.2.2
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [a85ef50]
+  - @agent-infra/browser@0.1.1
+  - @ui-tars/sdk@1.2.2
+
 ## 1.2.2-beta.3
 
 ### Patch Changes