+++ packages/ui-tars/operators/nut-js/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ui-tars/operator-nut-js
 
+## 1.2.2
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [a85ef50]
+  - @ui-tars/shared@1.2.2
+  - @ui-tars/sdk@1.2.2
+
 ## 1.2.2-beta.3
 
 ### Patch Changes