+++ packages/ui-tars/sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ui-tars/sdk
 
+## 1.2.2
+
+### Patch Changes
+
+- a85ef50: chore: publish
+- Updated dependencies [a85ef50]
+  - @ui-tars/action-parser@1.2.2
+  - @ui-tars/shared@1.2.2
+
 ## 1.2.2-beta.3
 
 ### Patch Changes