+++ packages/ui-tars/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ui-tars/shared
 
+## 1.2.2
+
+### Patch Changes
+
+- a85ef50: chore: publish
+
 ## 1.2.2-beta.3
 
 ## 1.2.2-beta.2